refactor: Reduce modification of Cassandra types & support arrays
This commit supports two pieces of functionality:
    1) It refactors the types within the connector to be
       as closely aligned to the node cassandra-driver
       types.

    2) It fixes an error when performing a POST with
       array values.

1) The types used in Cassandra are:
    • asci
    • bigint
    • blob
    • Boolean
    • counter
    • decimal
    • double
    • float
    • inet
    • int
    • text
    • timestamp
    • timeuuid
    • uuid
    • varchar
    • varint
    • list
    • map
    • set
    • tuple

Many of the types have directly related types in JavaScript.
Most of the transformations are handle by the cassandra-driver,
however, the driver also provides JavaScript classes to represent
Cassandra types that are more complex (uuid, Tuple, etc).

The types that are represented by JavaScript classess need to be
set up as custom Loopback types. The changes within this commit
tries to keep those class types as unaltered as possible.

TODO: add BigDecimal, Long, InetAddress, Integer, LocalDate,
and LocalTime.

2) The node cassandra-driver automatically handles the
transformation of arrays to the respective values within the
identified table a query is executing against (particularly
within update CQL statements). The default functionality
of the Loopback MySQL Connector converts array values to
JSON encoded strings. This is unecessary when using the
cassandra-driver